requirements.txt
alembic==1.0.5
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
funcsigs==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
pathlib2==2.3.2
Pillow==5.3.0
pluggy==0.8.0
psycopg2-binary==2.7.5
py==1.7.0
pytest==3.9.1
pytest-flask==0.13.0
python-dateutil==2.7.5
python-editor==1.0.3
requests==2.20.1
scandir==1.9.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24.1
waitress==1.1.0
Werkzeug==0.14.1
requests==2.20.1
